backend/pyproject.toml
[project]
name = "stablesoaps"
description = "Random code and notebooks for Stable Soaps"
version = "0.0.1"
authors = [
    {name = "fozziethebeat"},
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "anthropic",
    "jinja2",
    "jupyter",
    "Pillow",
    "pydantic",
    "python-dotenv",
    "sqids",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 80
